feat: refactor API models and endpoints for user and wallet management
- IBitGoApi.cs
  - Updated request and response model namespaces for PendingApprovals and Users.
  - Modified GetWalletAsync endpoint to include allTokens parameter.

- IBitGoApiService.cs
  - Updated request and response model namespaces for PendingApprovals and Users.

- Removed obsolete models related to PendingApproval and User login.
  - UpdatePendingApprovalModel.cs
  - LoginModel.cs
  - PendingApprovalModel.cs
  - PendingApprovalsListModel.cs
  - EntryModel.cs
  - TransferListModel.cs
  - TransferModel.cs
  - UserEmailModel.cs
  - UserModel.cs
  - UsernameModel.cs
  - CoinSpecificModel.cs
  - ReceiveAddressModel.cs
  - UserModel.cs
  - WalletListModel.cs
  - WalletModel.cs

- Added new models for PendingApprovals and Transfers.
  - UpdatePendingApprovalModel.cs
  - LoginModel.cs
  - PendingApprovalModel.cs
  - PendingApprovalsListModel.cs
  - EntryModel.cs
  - TransferListModel.cs
  - TransferModel.cs
  - UserEmailModel.cs
  - UserModel.cs
  - UsernameModel.cs
  - CoinSpecificModel.cs
  - ReceiveAddressModel.cs
  - UserModel.cs
  - WalletListModel.cs
  - WalletModel.cs
