+++ src/BitGo.Api.Client/Models/Responses/Users/UserModel.cs
@@ -1,4 +1,4 @@
-namespace BitGo.Api.Client.Models.Responses.User;
+namespace BitGo.Api.Client.Models.Responses.Users;
 
 public class UserModel
 {
